+++ Changelog.md
@@ -7,6 +7,8 @@
 - Fix Nullable settings in settings form (PlayerOptions stuff)
 - Fix issue where dungeonmaps would no longer draw after changing game resolution
 - Fix /ub igp/give commands to be non greedy when matching profile (fixes /ub ig test.utl to char with to in the name)
+- Add expressions: wobjectfindall[], wobjectfindallinventory[], wobjectfindalllandscape[], wobjectfindallbyobjectclass[objectclass], wobjectfindallbytemplatetype[templatetype], wobjectfindallbynamerx[namerx], wobjectfindallinventorybyobjectclass[objectclass], wobjectfindallinventorybytemplatetype[templatetype], wobjectfindallinventorybynamerx[namerd], wobjectfindalllandscapebyobjectclass[objectclass], wobjectfindalllandscapebytemplatetype[templatetype], wobjectfindalllandscapebynamerx[namerx]
+- Add expressions: listmap[list, expression], listreduce[list, expression], listfilter[list, expression]
 
 ## 0.2.5 (2021-10-06) [UtilityBeltInstaller-0.2.5.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/dd7285b144ef9327bf8041c64841c22f/UtilityBeltInstaller-0.2.5.exe)
 - Add expressions: wobjectgetdoubleprop[], wobjectgetboolprop[], wobjectgetstringprop[], wobjecthasdata[], wobjectrequestdata[]